Add Windows IOCP poller and cross-platform socket/pipe support
- Implement IOCP-based IPoller in src/win_poller.cpp (CreateIoCompletionPort,
  GetQueuedCompletionStatus, overlapped WSARecv/WSASend for level-triggered events)
- Add poll_fd_t (int on Unix, intptr_t on Windows) and optional IPoller::wakeup()
- Event loop: Windows uses make_poller(this) and PostQueuedCompletionStatus wakeup
- CMake: detect Windows, add win_poller.cpp, link ws2_32, nghttp2 CONFIG on Windows
- Sockets: WSASocket(WSA_FLAG_OVERLAPPED), closesocket, recv/send, nonblock via
  ioctlsocket(FIONBIO) on Windows; tcp/udp and headers use poll_fd_t
- GitHub Actions: add build-windows job (windows-latest, vcpkg openssl+nghttp2, MSVC)

+++ src/http_server.cpp
@@ -313,7 +313,7 @@ Task<void> run_http_server(EventLoop& loop, HttpServerOptions options) {
                 client.close();
                 continue;
             }
-            SSL_set_fd(ssl, client.fd());
+            SSL_set_fd(ssl, static_cast<int>(client.fd()));
             TlsSocket tls(std::move(client), ssl);
             std::error_code err = co_await tls.async_handshake_server(loop);
             if (err) {
